Add a new class to handle api requests
