+++ iamctl/pkg/applications/applicationUtils.go
@@ -19,12 +19,9 @@
 package applications
 
 import (
-	"bytes"
-	"crypto/tls"
 	"encoding/json"
 	"io/ioutil"
 	"log"
-	"net/http"
 	"os"
 	"text/tabwriter"
 
@@ -56,22 +53,11 @@ func getDeployedAppNames() []string {
 
 func getAppList() (spIdList []Application) {
 
-	var APPURL = utils.SERVER_CONFIGS.ServerUrl + "/t/" + utils.SERVER_CONFIGS.TenantDomain + "/api/server/v1/applications"
 	var list AppList
-
-	http.DefaultTransport.(*http.Transport).TLSClientConfig = &tls.Config{InsecureSkipVerify: true}
-
-	req, _ := http.NewRequest("GET", APPURL, bytes.NewBuffer(nil))
-	req.Header.Set("Authorization", "Bearer "+utils.SERVER_CONFIGS.Token)
-	req.Header.Set("accept", "*/*")
-	defer req.Body.Close()
-
-	httpClient := &http.Client{}
-	resp, err := httpClient.Do(req)
+	resp, err := utils.SendGetListRequest(utils.APPLICATIONS)
 	if err != nil {
-		log.Fatalln(err)
+		log.Println("Error while retrieving application list", err)
 	}
-
 	defer resp.Body.Close()
 
 	statusCode := resp.StatusCode
